Enable language change in the middle of test and after results
