+++ lib/modules/home/home_page_controller.dart
@@ -8,6 +8,7 @@ import 'package:palumba_eu/data/repositories/local/local_data_repository.dart';
 import 'package:palumba_eu/modules/onboarding/onboarding_controller.dart';
 import 'package:palumba_eu/modules/results/results_controller.dart';
 import 'package:palumba_eu/modules/settings/settings_page_controller.dart';
+import 'package:palumba_eu/modules/statments/statements_screen_controller.dart';
 import 'package:palumba_eu/utils/managers/user_manager.dart';
 import 'package:palumba_eu/utils/string_utils.dart';
 import 'package:palumba_eu/utils/utils.dart';
@@ -80,7 +81,7 @@ class HomePageController extends GetxController {
 
   void backToResultsOrTest() {
     if (isTestRunning) {
-      Get.back();
+      Get.toNamed(StatementsController.route);
       return;
     }
     Get.toNamed(ResultsController.route, arguments: {
