+++ lib/utils/managers/language_manager.dart
@@ -1,9 +1,11 @@
 import 'package:get/get.dart';
+import 'package:palumba_eu/data/repositories/remote/data_repository.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'dart:ui';
 
 class LanguageManager {
   static late SharedPreferences sharedPreferences;
+  static final DataRepository _dataRepository = Get.find<DataRepository>();
 
   static String currentLanguage = 'en';
   static const String keyLanguage = 'language';
@@ -37,6 +39,7 @@ class LanguageManager {
     currentLanguage = languageCode;
     setSavedLanguage(languageCode);
     Get.updateLocale(Locale(languageCode));
+    _dataRepository.fetchTranslatedData();
   }
 
   static String getApiLanguage() {
